backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
asyncpg==0.29.0
pydantic==2.7.4
pydantic-settings==2.3.4
google-generativeai==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.5.1
httpx==0.27.0
python-multipart==0.0.9
pypdf2==3.0.1
python-docx==1.1.2
spacy==3.7.5
web3==6.19.0
python-dotenv==1.0.1
alembic==1.13.2
